Decode NMEA Messages

National Marine Electronics Association (NMEA) messages are comma separated value lines of text that start with sender or "talker" code and the sentence type.

ais_decode_nmea(
  msgs,
  data_type = c("auto", "file", "strings", "text"),
  as_df = maritime_as_df(),
  as_sf = maritime_as_sf(),
  as_tibble = maritime_as_tibble(),
  verbose = maritime_verbose()
)

Arguments

msgs
character()

NMEA messages

data_type
character(1L)

how to interpret x.

one of "auto", "file", "strings", or "text"

"auto": let ais_decode() decide

"file": treat x as a file path

"strings": treat x as a vector with each element containing a single AIS message

"text": treat x as a single character containing multiple AIS messages

as_df
logical(1L), default: maritime_as_df()

Whether the parsed content should be transformed to a data frame.

maritime_as_df() defaults to TRUE.

as_sf
logical(1L), default: maritime_as_sf()

Whether an sf (sf::st_sf()) should be returned.

Ignored if the {sf} package is not installed.

maritime_as_sf() defaults to TRUE.

as_tibble
logical(1L), default: maritime_as_tibble()

Whether a tibble::tibble() should be returned.

Ignored if the {tibble} package is not installed.

maritime_as_tibble() defaults to TRUE.

verbose
logical(1L), default: maritime_verbose()

Whether print status and diagnostic messages.

maritime_verbose() defaults to TRUE.

Details

ais_decode_nmea decodes NMEA messages (currently 1, 2, and 3) contained in files or character vectors.

Examples

nmea_file <- example_nmea_file()
nmea_lines <- readLines(nmea_file, warn = FALSE)

nmea_lines[1:10]

#>  [1] "!AIVDM,1,1,,A,13u?etPv2;0n:dDPwUM1U1Cb069D,0*23"                                            
#>  [2] ",1541030400"                                                                                
#>  [3] ",1541030400"                                                                                
#>  [4] "!ABVDM,1,1,7,B,H3uJVD4NDBE5847C62ppnk0@6120,0*4F,1541030311,1541030400"                     
#>  [5] "!BSVDM,2,1,9,A,53KPtl01pv8`tTtR22050thhtr2222222222220l1`A916Vh0;20DPSmD`0D,0*21,1541030311"
#>  [6] "!BSVDM,2,2,9,A,ljDm0CPiH80,2*21,1541030311,1541030400"                                      
#>  [7] "!BSVDM,1,1,,B,147lQH00001f>MtQtr3a92nr00Sa,0*0B,1541030311,1541030400"                      
#>  [8] "!AIVDM,1,1,0,A,H4eG>Gh5@h4q@T>1<uDp@000000,2*25,1541030311,1541030400"                      
#>  [9] "!AIVDM,1,1,,A,33e3Dn500012NAtDR:WV5H<r0E0r,0*0E,1541030311,1541030400"                      
#> [10] "!BSVDM,1,1,,A,H3n2?6Q=H4pDr22222222222220,2*47,1541030311,1541030400"                       

# decoding a file =====================================================================
ais_decode_nmea(nmea_file)

#> - `data_type` detected: file

#> Simple feature collection with 27 features and 26 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -72.54038 ymin: 14.68399 xmax: 25.5757 ymax: 69.31088
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> # A tibble: 27 x 27
#>    message_id repeat_indicator   mmsi rot_over_range     rot    sog
#>         <int>            <int>  <int> <lgl>            <dbl>  <dbl>
#>  1          1                0 2.66e8 FALSE            -2.86 13.9  
#>  2          1                0 2.77e8 FALSE             0     0    
#>  3          3                0 2.49e8 FALSE             0     0    
#>  4          3                0 2.57e8 FALSE             0     0    
#>  5          1                0 2.45e8 TRUE           -731.    0    
#>  6          1                0 2.06e8 TRUE           -731.    0    
#>  7          3                1 2.31e8 FALSE             0     0    
#>  8          1                0 6.63e8 TRUE           -731.    0.100
#>  9          1                0 2.45e8 TRUE           -731.    0    
#> 10          1                1 2.30e8 FALSE             0     0    
#> # … with 17 more rows, and 21 more variables: position_accuracy <int>,
#> #   lng_deg <dbl>, lat_deg <dbl>, cog <dbl>, true_heading <int>,
#> #   timestamp <int>, special_manoeuvre <int>, spare <int>, raim <lgl>,
#> #   sync_state <int>, slot_timeout <int>, received_stations <int>,
#> #   slot_number <int>, utc_hour <int>, utc_min <int>, utc_spare <int>,
#> #   slot_offset <int>, slot_increment <int>, slots_to_allocate <int>,
#> #   keep_flag <lgl>, geometry <POINT [°]>